% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDC-package.R
\docType{data}
\name{amphibians_chi}
\alias{amphibians_chi}
\title{amphibian range maps for the extent of Chiquitos}
\format{
A \code{data.frame} with 23 observations and 26 variables.
}
\source{
This data has been obtained from: 
\itemize{\item \url{https://www.iucnredlist.org/resources/spatial-data-download}}
}
\usage{
data(amphibians_chi)
}
\description{
IUCN Amphibian ranges with Chiquitos, Bolivia
}
\details{
shapefiles of the global IUCN ranges of all amphibian species that intersect with the extent of Chiquitos
}
