% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDC-package.R
\docType{data}
\name{biotime_bol}
\alias{biotime_bol}
\title{bioTime data for Bolivia}
\format{
A \code{data.frame} with 206 observation and 15 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://biotime.st-andrews.ac.uk/download_form.php?dl=csv_download}}
}
\usage{
data(biotime_bol)
}
\description{
biodiversity time series trends from the BioTime database for Bolivia
}
\details{
Biodiversity time series trends extracted from the BioTime database for the 
extent of Bolivia. For Chiquitos and Bolivia no data is available, but some data is available for Chile (which is provided here).
}
\references{
Dornelas M, Antao LH, Moyes F, Bates, AE, Magurran, AE, et al. BioTIME: 
A database of biodiversity time series for the Anthropocene. 
Global Ecol Biogeogr. 2018; 27:760-786. \url{https://doi.org/10.1111/geb.12729}
}
