% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDC-package.R
\docType{data}
\name{bolivia}
\alias{bolivia}
\title{administrative boundary of bolivia}
\format{
A \code{data.frame} with 1 observation and 3 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://www.gadm.org/data.html}}
}
\usage{
data(bolivia)
}
\description{
GADM outline of Bolivia
}
\details{
Administrative boundary of Bolivia derived from GADM data.
}
