% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDC-package.R
\docType{data}
\name{cci_chi}
\alias{cci_chi}
\title{CCI land cover data of chiquitos}
\format{
A \code{data.frame} with 1122489 observations and 31 variables.
}
\usage{
data(cci_chi)
}
\description{
CCI land cover data of Chiquitos, Bolivia
}
\details{
CCI land cover data of Chiquitos, Bolivia at a resolution of 300m. 
CCI Land cover are global land cover maps at 300m spatial resolution based on MERIS and SPOT satellite data. 
Data includes annual information from 1992 - 2022.
}
