% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDC-package.R
\name{CDC package}
\alias{CDC package}
\alias{CDCpackage}
\title{CDC: Chiquitos Data Cube}
\description{
R package to access various data sources, i.e. elevation, climate and land-use data, for the Chiquitos region in Bolivia.
}
\details{
CDC: Chiquitos Data Cube
}
\author{
MattB-SF
}
\keyword{package}
