% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDC-package.R
\docType{data}
\name{chi_pop_1km}
\alias{chi_pop_1km}
\title{Population density of Chiquitos}
\format{
A \code{data.frame} with 70552 observations and 5 variables.
}
\usage{
data(bav_pop_1km)
}
\description{
Population density in Chiquitos, Bolivia
}
\details{
Population density of Chiquitos at a resolution of 1km2
}
