% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDC-package.R
\docType{data}
\name{chiquitos}
\alias{chiquitos}
\title{administrative boundary of chiquitos}
\format{
A \code{data.frame} with 1 observation and 14 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://www.gadm.org/data.html}}
}
\usage{
data(chiquitos)
}
\description{
GADM outline of Chiquitos, Bolivia
}
\details{
Administrative boundary of Chiquitos, Bolivia derived from GADM data.
}
