% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDC-package.R
\docType{data}
\name{pa_chi}
\alias{pa_chi}
\title{protected area data of chiquitos}
\format{
A \code{data.frame} with 8 observations and 31 variables.
}
\source{
Cite this dataset as: \itemize{\item UNEP-WCMC and IUCN (2019), Protected Planet: 
The World Database on Protected Areas (WDPA) On-line, July/2019, Cambridge, UK: UNEP-WCMC and IUCN. 
Available at: \url{https://protectedplanet.net.}.}
}
\usage{
data(pa_chi)
}
\description{
IUCN protected area data of Chiquitos, Bolivia
}
\details{
Protected area data of Chiquitos, Bolivia derived from the World Database on Protected Areas
}
