% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDC-package.R
\docType{data}
\name{santa_cruz}
\alias{santa_cruz}
\title{administrative boundary of santa cruz}
\format{
A \code{data.frame} with 1 observation and 12 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://www.gadm.org/data.html}}
}
\usage{
data(santa_cruz)
}
\description{
GADM outline of Santa Cruz, Bolivia
}
\details{
Administrative boundary of Santa Cruz, Bolivia derived from GADM data.
}
