% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDC-package.R
\docType{data}
\name{wc2.1_30s_chi}
\alias{wc2.1_30s_chi}
\title{worldclim data of chiquitos}
\format{
A \code{data.frame} with 92600 observation and 93 variables.
}
\source{
Cite this dataset as: \itemize{\item Hijmans, R.J., S.E. Cameron, J.L. Parra, P.G. Jones and A. Jarvis, 2005. 
Very high resolution interpolated climate surfaces for global land areas. 
International Journal of Climatology 25: 1965-1978.}
This data has been downloaded from: \itemize{\item \url{https://worldclim.org/version1}}
}
\usage{
data(wc2.1_30s_chi)
}
\description{
Current worldclim temperature, precipitation and bioclim data of Chiquitos, Bolivia
}
\details{
Current temperature, precipitation, solar radiation, wind and bioclimatic data of Chiquitos, Bolivia derived 
from 30 seconds Worldclim v2.1. Data is derived from interpolations of observed data representative of xx - xx.
}
